исправлена верстка экрана создания трекера. Исправлено условие placeholder

+++ Tracker/Trackers/Schedule/ScheduleStructure.swift
@@ -38,16 +38,16 @@ enum Weekday: Int, CaseIterable {
         }
     }
     
-    static func calculateScheduleValue(for schedule: [Weekday]) -> Int32 {
-        var scheduleValue: Int32 = 0
+    static func calculateScheduleValue(for schedule: [Weekday]) -> Int16 {
+        var scheduleValue: Int16 = 0
         for day in schedule {
-            let dayRawValue = Int32 (1 << day.rawValue)
+            let dayRawValue = Int16 (1 << day.rawValue)
             scheduleValue |= dayRawValue
         }
         return scheduleValue
     }
     
-   static func calculateScheduleArray(from value: Int32) -> [Weekday] {
+   static func calculateScheduleArray(from value: Int16) -> [Weekday] {
         var schedule: [Weekday] = []
         for day in Weekday.allCases {
             if value & (1 << day.rawValue) != 0 {
